20220527 프로그래머스 자바 12969 풀이

+++ baekjoon/src/com/potatomeme/may/week_4th/p_solution_12954.java
@@ -3,8 +3,8 @@ package com.potatomeme.may.week_4th;
 public class p_solution_12954 {
     public long[] solution(int x, int n) {
         long[] answer = {};
-        for(int i = 0;i<n;i++){
-            answer[i] = x*(i+1);
+        for (int i = 0; i < n; i++) {
+            answer[i] = (long) x * (i + 1);
         }
         return answer;
     }
